tests structured alike the classes to test.
